linear combination
Now have three f values (number of conflicts, distance traveled, and
time spent in air). Note number of conflicts is not the global number
of conflicts but rather represents the conflicts that are counted for
or against an agent depending on it’s learning behavior. the program
does currently build but the EA and other calculations need to be
re-verified.